+++ dune/common/fmatrix.hh
@@ -14,6 +14,7 @@
 #include <dune/common/fvector.hh>
 #include <dune/common/densematrix.hh>
 #include <dune/common/precision.hh>
+#include <dune/common/promotiontraits.hh>
 #include <dune/common/typetraits.hh>
 
 namespace Dune
@@ -121,6 +122,34 @@ namespace Dune
     //! no copy assignment from FieldMatrix of different size
     template <typename T, int rows, int cols>
     FieldMatrix& operator=(FieldMatrix<T,rows,cols> const&) = delete;
+
+    //! vector space addition -- two-argument version
+    template <class OtherScalar>
+    friend auto operator+ ( const FieldMatrix& matrixA,
+                            const FieldMatrix<OtherScalar,ROWS,COLS>& matrixB)
+    {
+      FieldMatrix<typename PromotionTraits<K,OtherScalar>::PromotedType,ROWS,COLS> result;
+
+      for (size_type i = 0; i < ROWS; ++i)
+        for (size_type j = 0; j < COLS; ++j)
+          result[i][j] = matrixA[i][j] + matrixB[i][j];
+
+      return result;
+    }
+
+    //! vector space subtraction -- two-argument version
+    template <class OtherScalar>
+    friend auto operator- ( const FieldMatrix& matrixA,
+                            const FieldMatrix<OtherScalar,ROWS,COLS>& matrixB)
+    {
+      FieldMatrix<typename PromotionTraits<K,OtherScalar>::PromotedType,ROWS,COLS> result;
+
+      for (size_type i = 0; i < ROWS; ++i)
+        for (size_type j = 0; j < COLS; ++j)
+          result[i][j] = matrixA[i][j] - matrixB[i][j];
+
+      return result;
+    }
 
     //! Multiplies M from the left to this matrix, this matrix is not modified
     template<int l>
@@ -248,6 +277,58 @@ namespace Dune
     }
 
     using Base::operator=;
+
+    //! vector space addition -- two-argument version
+    template <class OtherScalar>
+    friend auto operator+ ( const FieldMatrix& matrixA,
+                            const FieldMatrix<OtherScalar,1,1>& matrixB)
+    {
+      return FieldMatrix<typename PromotionTraits<K,OtherScalar>::PromotedType,1,1>{matrixA[0][0] + matrixB[0][0]};
+    }
+
+    //! Binary addition when treating FieldMatrix<K,1,1> like K
+    template <class Scalar,
+              std::enable_if_t<IsNumber<Scalar>::value, int> = 0>
+    friend auto operator+ ( const FieldMatrix& matrix,
+                            const Scalar& scalar)
+    {
+      return FieldMatrix<typename PromotionTraits<K,Scalar>::PromotedType,1,1>{matrix[0][0] + scalar};
+    }
+
+    //! Binary addition when treating FieldMatrix<K,1,1> like K
+    template <class Scalar,
+              std::enable_if_t<IsNumber<Scalar>::value, int> = 0>
+    friend auto operator+ ( const Scalar& scalar,
+                            const FieldMatrix& matrix)
+    {
+      return FieldMatrix<typename PromotionTraits<Scalar,K>::PromotedType,1,1>{scalar + matrix[0][0]};
+    }
+
+    //! vector space subtraction -- two-argument version
+    template <class OtherScalar>
+    friend auto operator- ( const FieldMatrix& matrixA,
+                            const FieldMatrix<OtherScalar,1,1>& matrixB)
+    {
+      return FieldMatrix<typename PromotionTraits<K,OtherScalar>::PromotedType,1,1>{matrixA[0][0] - matrixB[0][0]};
+    }
+
+    //! Binary subtraction when treating FieldMatrix<K,1,1> like K
+    template <class Scalar,
+              std::enable_if_t<IsNumber<Scalar>::value, int> = 0>
+    friend auto operator- ( const FieldMatrix& matrix,
+                            const Scalar& scalar)
+    {
+      return FieldMatrix<typename PromotionTraits<K,Scalar>::PromotedType,1,1>{matrix[0][0] - scalar};
+    }
+
+    //! Binary subtraction when treating FieldMatrix<K,1,1> like K
+    template <class Scalar,
+              std::enable_if_t<IsNumber<Scalar>::value, int> = 0>
+    friend auto operator- ( const Scalar& scalar,
+                            const FieldMatrix& matrix)
+    {
+      return FieldMatrix<typename PromotionTraits<Scalar,K>::PromotedType,1,1>{scalar - matrix[0][0]};
+    }
 
     //===== solve