+++ CHANGELOG.md
@@ -34,6 +34,13 @@
   - Vector = Scalar * Vector
   - Vector = Vector * Scalar
   - Vector = Vector / Scalar
+  Correspondingly, the `FieldMatrix` class now has
+  - Matrix = Matrix + Matrix
+  - Matrix = Matrix - Matrix
+  Note that the operators
+  - Vector = Vector + Vector
+  - Vector = Vector - Vector
+  have been introduced earlier.
 
 - There is now (finally!) a method `power` in the file `math.hh` that computes
   powers with an integer exponent, and is usable in compile-time expressions.